refactor(store-core): extract web-task + search handlers into handlers/web-task.ts (audit P2-3)
Move the web-task + search family (handleWebTaskUpsert, applyWebTaskUpsert,
handleWebTaskError, handleWebTaskList, handleWebFeatureStatus, handleSearchResults
+ their payload interfaces) out of the handlers/index.ts god-module into
handlers/web-task.ts, behind the barrel's export * re-export so the public
surface is unchanged.

Drop the now-unused WebTask and SearchResult type imports from index.ts (they
were only used by the moved handlers).

Pure move — body byte-identical to origin/main. index.ts 2209 -> 1995 lines.
